CoverageTransformer class
