+++ lib/SparceCoverageCollector.js
@@ -41,7 +41,7 @@ class SparceCoverageCollector {
   }
 
 
-  getFilesCoverage() {
+  getFinalCoverage() {
     return this.srcCoverage;
   }
 
@@ -113,4 +113,4 @@ class SparceCoverageCollector {
   }
 }
 
-module.exports = SparceCoverageCollector;
+module.exports.SparceCoverageCollector = SparceCoverageCollector;
